move checkbox stories to docs
